parsers/llm-parser/requirements.txt
requests
python-dotenv
python-docx
supabase
pywin32
